# swagger.yaml
openapi: 3.0.0

info:
  title: Chat App API
  version: 1.0.0
  description: API documentation for the chat application

servers:
  - url: http://localhost:3000

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

paths:
  /auth/login:
    post:
      summary: Login a user
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
              properties:
                username:
                  type: string
                  example: user123
                password:
                  type: string
                  example: secretpass
      responses:
        "200":
          description: Successful login with JWT token
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        "401":
          description: Invalid username or password
        "500":
          description: Server error

  /auth/register:
    post:
      summary: Register a new user
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
                - name
                - email
              properties:
                username:
                  type: string
                  example: user123
                password:
                  type: string
                  example: secretpass
                name:
                  type: string
                  example: John
                surname:
                  type: string
                  example: Doe
                email:
                  type: string
                  example: john@example.com
      responses:
        "201":
          description: User registered successfully
        "400":
          description: Username already exists or Email already in use
        "500":
          description: Server error

  /auth/me:
    get:
      summary: Get current user profile
      tags: [Auth]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user info from token
          content:
            application/json:
              schema:
                type: object
                properties:
                  userId:
                    type: string
                  username:
                    type: string
                  name:
                    type: string
                  surname:
                    type: string
                  email:
                    type: string
        "401":
          description: Unauthorized

  /auth/users:
    get:
      summary: Get all users except current user
      tags: [Auth]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    _id:
                      type: string
                    username:
                      type: string
                    name:
                      type: string
                    surname:
                      type: string
                    email:
                      type: string
        "401":
          description: Unauthorized
        "500":
          description: Server error

  /chat:
    post:
      summary: Create a new chat
      tags: [Chat]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - members
                - name
                - isGroup
              properties:
                members:
                  type: array
                  items:
                    type: string
                    description: MongoDB ObjectId
                  example:
                    ["67a123fbc9123e88d12f456a", "67b456fbc9123e88d12f789b"]
                name:
                  type: string
                  example: Study Group
                isGroup:
                  type: boolean
                  example: true
      responses:
        "201":
          description: Chat created successfully
        "400":
          description: Bad request
        "500":
          description: Server error

    get:
      summary: Get all chats for the current user
      tags: [Chat]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of chats
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    _id:
                      type: string
                    name:
                      type: string
                    isGroup:
                      type: boolean
                    members:
                      type: array
                      items:
                        type: object
                    lastMessage:
                      type: string
                    createdAt:
                      type: string
                    updatedAt:
                      type: string
        "401":
          description: Unauthorized
        "500":
          description: Server error

  /chat/{id}/messages:
    get:
      summary: Get paginated messages for a chat
      tags: [Chat]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: Chat MongoDB ObjectId
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 30
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    _id:
                      type: string
                    chat:
                      type: string
                    sender:
                      type: object
                    content:
                      type: string
                    createdAt:
                      type: string
        "401":
          description: Unauthorized
        "500":
          description: Server error